requirements.txt
python-telegram-bot
python-dotenv
openai
requests
langchain
langchain_community
langchain-openai
faiss-cpu
pypdf